+++ src/ui/components/seekbar/SeekBar.tsx
@@ -74,7 +74,10 @@ export class SeekBar extends PureComponent<SeekBarProps, SeekBarState> {
     }
   };
   private _onLoadedMetadata = (event: LoadedMetadataEvent) => this.setState({ duration: event.duration });
-  private _onDurationChange = (event: DurationChangeEvent) => this.setState({ duration: event.duration });
+  private _onDurationChange = (event: DurationChangeEvent) => {
+    const player = (this.context as UiContext).player;
+    this.setState({ duration: event.duration, seekable: player.seekable });
+  }
   private _onProgress = (event: ProgressEvent) => this.setState({ seekable: event.seekable });
 
   private _onSlidingStart = (value: number) => {